Saat Güncellemesi
RecyclerView'a bağlı adaptördeki saat formatı ve detay sayfasındaki saat formatı düzeltildi

+++ app/src/main/java/com/tugrulbo/kotlinflightapp/model/Arrival.kt
@@ -11,7 +11,7 @@ data class Arrival(
     @SerializedName("airport")
     var airport: String?,
     @SerializedName("delay")
-    var delay: Int?,
+    var delay: Float?,
     @SerializedName("estimated")
     var estimated: String?,
     @SerializedName("gate")
